backend/requirements.txt
"fastapi[standard]"
sqlalchemy 
psycopg2-binary
python-dotenv
os